Move ErrorListener to src/main/java to expose it to consumer.
